1. Replace size_t with uint32_t to fix 64/32 bit compatibility issues 2. Fixed memory bug related to data array and bulk loading for RTree.

+++ include/LineSegment.h
@@ -27,7 +27,7 @@ namespace SpatialIndex
 	{
 	public:
 		LineSegment();
-		LineSegment(const double* startPoint, const double* endPoint, size_t dimension);
+		LineSegment(const double* startPoint, const double* endPoint, uint32_t dimension);
 		LineSegment(const Point& startPoint, const Point& endPoint);
 		LineSegment(const LineSegment& l);
 		virtual ~LineSegment();
@@ -43,9 +43,9 @@ namespace SpatialIndex
 		//
 		// ISerializable interface
 		//
-		virtual size_t getByteArraySize();
+		virtual uint32_t getByteArraySize();
 		virtual void loadFromByteArray(const byte* data);
-		virtual void storeToByteArray(byte** data, size_t& length);
+		virtual void storeToByteArray(byte** data, uint32_t& length);
 
 		//
 		// IShape interface
@@ -54,7 +54,7 @@ namespace SpatialIndex
 		virtual bool containsShape(const IShape& in) const;
 		virtual bool touchesShape(const IShape& in) const;
 		virtual void getCenter(Point& out) const;
-		virtual size_t getDimension() const;
+		virtual uint32_t getDimension() const;
 		virtual void getMBR(Region& out) const;
 		virtual double getArea() const;
 		virtual double getMinimumDistance(const IShape& in) const;
@@ -65,11 +65,11 @@ namespace SpatialIndex
 		virtual double getRelativeMaximumDistance(const Region& r) const;
 		virtual double getAngleOfPerpendicularRay();
 
-		virtual void makeInfinite(size_t dimension);
-		virtual void makeDimension(size_t dimension);
+		virtual void makeInfinite(uint32_t dimension);
+		virtual void makeDimension(uint32_t dimension);
 
 	public:
-		size_t m_dimension;
+		uint32_t m_dimension;
 		double* m_pStartPoint;
 		double* m_pEndPoint;
